+++ tests/lossless_coding/test__fixed_to_variable_codes.py
@@ -10,14 +10,18 @@ from .util import deterministic_pmf, random_pmf, shuffle_pmf
 
 @pytest.fixture(
     params=[
-        partial(komm.HuffmanCode, policy="high"),
-        partial(komm.HuffmanCode, policy="low"),
+        partial(komm.HuffmanCode, policy="high", assignment="tree"),
+        partial(komm.HuffmanCode, policy="low", assignment="tree"),
+        partial(komm.HuffmanCode, policy="high", assignment="canonical"),
+        partial(komm.HuffmanCode, policy="low", assignment="canonical"),
         komm.ShannonCode,
         komm.FanoCode,
     ],
     ids=[
-        "HuffmanCode-high",
-        "HuffmanCode-low",
+        "HuffmanCode-high-tree",
+        "HuffmanCode-low-tree",
+        "HuffmanCode-high-canonical",
+        "HuffmanCode-low-canonical",
         "ShannonCode",
         "FanoCode",
     ],